core: major cleanup in color- and emoticon-selectors code
Get rid of unnecessary slots ColorSelector::alignTo() and
EmoticonSelector::alignTo() and move their body to the corresponding
constructors. Because of this it was necessary to pass pointers to the
activating widgets to the constructors.
Make EmoticonSelector not know about its calling widget. Instead add a
signal to select an emoticon.
Get rid of unnecessary ColorSelector::aboutToClose() signal. Also remove
ColorSelector::closeEvent() since it's not needed anymore.

Commited also some cosmetic changes, mostly in formatting.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>