idle/x11: Open and use own Display rather than QApplication's
Rationale behind this change is the fact that XScreenSaver extension
registers its close function in the given Display, for they to be called
in XCloseDisplay(). However, since the idle can be a dynamically loaded
module, it is unloaded along with -lXss (which kadu-core doesn't link
to) before we want to actually close QApplication's Display. So, when
doing XCloseDisplay() called from qt_cleanup(), libX11 will try to
execute a function provided earlier by XScreenSaver, which no longer
stays in memory. In such case, bad things will happen (actually, the
stack gets corrupted and we get a segfault).

Note that currently Kadu will never call QApplication's desctructor and
the described trouble doesn't take place. Though, now we will be able to
properly destroy QApplication instance.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>