gui: Rewrite tab handling in ConfigSection
Instead of creating new regular widget or new tab widget depending on
how many widgets are to be shown, use one tab widget with custom method
that lets us hide the tab bar. It simplifies things a lot.

The behavioral change is that it fixes this bug:
  1. Load autoway.
  2. Open configuration window.
  3. Unload autoaway.
  4. "Status" section in configuration window now has no widget visible,
     while it should have one.

This bug was hard to fix with previous code design (or I missed
something).

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>