Use Account::data() rather than Account::statusContainer() in some places
In some places we want to get status information specific to a given
account, even in simple mode. Amongst others, it fixes a bug in
ContactDataExtractor that caused incorrect status icons to be displayed
in simple mode. The issue was spotted by White Eagle.

Also, document Account::statusContainer() to avoid this kind of bugs in
future.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>