gui, chat: Introduce new fake kaduimg protocol for handling relative paths for saved images from chats (#1635)
The idea is to deal everywhere where possible with relative paths for
saved images (actually it's only its file name) and pass it as path to
kaduimg:// protocol which is handled by ChatViewNetworkAccessManager.

Currently, a hack is needed for sent images, because as of now we aren't
saving them anywhere - we use their full system paths.

http://kadu.net/mantis/view.php?id=1635

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>