Fix Shared objects not ever being destroyed (#1925)
The patch basically contains of three steps:

1) Remove data()->ref.ref() calls from SharedBase c-tors. This allows
Shared objects to be actually ever deleted, otherwise reference count
was always at least 1.

2) Add ref.ref() calls to Shared d-tors to prevent infinite recursion.
(Theoretically it'd be suffiecient to add those only to one or two
classes but I decided to do it for all just in case). BTW, ref.deref()
calls aren't necessary since the object will be destroyed anyway if we
are in its d-tor.

3) Add guards to some methods during which execution "this" can be
deleted, causing a crash.

It was a really big issue. I hope I covered all possible crashed.
Everything is described in details in the comments.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>