gadu: Use QString rather than QByteArray when reading received message formatting
It's needed because text position in formatting data treats Unicode
characters as one character, not as two bytes like we did before.

It also fixes message encoding in case we don't even try to read
formatting (there was a bug in that code path not creating a QString
with QString::fromUtf8()). It was the case when the sender wasn't on our
contact list and also for Blip (I'm not sure why).

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>