gui: Fix crash on Windows after right clicking on a toolbar
Apparently mouse events are delivered differently on Windows than on X11
and we were refreshing toolbars after right clicking on them. And it was
causing a crash because we were trying to execute a menu on that
toolbar (note that refreshing is implemented as destroying a toolbar and
then creating a new one).

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>