gui/actions: Fix removing ActionDataSource from MappedActions after receiving destroyed() signal from an Action in ActionDescription class in certain cases
This fixes crash when using CustomInputMenuManager.

Note that Action::aboutToBeDestroyed() signal had to be introduced to be
able to retrieve ActionDataSource from the Action object before it is
destroyed. QObject::destroyed() is emitted in ~QObject(), so it's too
late as the object at that time is only a QObject and anything more is
already destroyed.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>