plugins: Correctly import 0.9 configuration
First, we were setting state of never loaded plugins to
PluginStateDisabled, what is incorrect (correct state is the default,
i.e., PluginStateNew).

Second, the new ensureLoadedAtLeastOnce() was not equivalent to the old
one and with the current system it just couldn't fix anything: if
a loaded-by-default-plugin's state is PluginStateNew, it will for sure
be loaded at least once by the new class. And if its state was really
PluginStateNew, it means that it was never loaded (probably due to a bug
in a not fixed version from the 0.9.x series), so the state must not be
changed until it will be actually loaded.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>