Do not store ShareBase objects in Shared objects
Thanks to this we do not need to include "sth1.h" in "sth2-shared.h"
which significantly reduces includes burden and allows to always include
"sth3-shared.h" in "sth3.h".

This change also makes Shared classes compatible with MSVC. It turns out
that MSVC cannot instantiate a template if it doesn't know definition of
a class that's passed to it as a template parameter.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>