gui: rework of ActionsComboBox
This is rather big change with all comboboxes in Kadu...
Short summary: activatedSlots and currentIndexChangedSlots are now only
in ActionsComboBox. This class handles all actions calling triggers when
activated (not in curretIndexChanged).

If action has boolean data value of true then it is "unselectable" action
and before triggering combobox state is restored to previous, so this
action will never be selected.

valueChanged method replaces old currentIndexChangedSlots that called
old ActionComboBox::currentIndexChangedSlots method.

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>