gui: Fix crash when the session manager forces Kadu to close and KaduWindow was visible
In that case KaduWindow accepted close event and deleted itself. And
Core::quit() was called after it, so when plugins were being unloaded
and they were unregistering actions etc. from KaduWindow (assuming it is
not null), KaduWindow was already null and it crashed.

DeleteOnClose is useless here anyway, we really shouldn't delete
KaduWindow before Core::quit(), and Core::quit() is always called from
KaduWindow::closeEvent() when it gets accepted.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>