gadu: Always recheck which socket notifiers to enable aftet libgadu might set GG_CHECK_WRITE
Description of the bug:
Say there is nothing interesting hapenning in a moment. So we have
ReadNotifier enabled but WriteNotifier is obviously off. Now we tell
libgadu to send something. Unfortunately kernel fails to send the entire
packet at once, say because it is a 15,000 characters long message or
the system is simply under heavy load. libgadu does what it should, i.e.
sets the GG_CHECK_WRITE flag to 1. Though, we do not enable
WriteNotifier immediately. Actually we will enable it only when
something will happen on ReadNotifier, which may not be the case even
until the server considers us dead and breaks connection. It is because
libgadu waits until we tell it that it is worth to try to send the rest
of the packet through calling gg_watch_fd(), and until we do it, it does
not send any more data, and only appends it to its internal buffer
instead. Thus, we cannot send anything and probably server will
eventually break connection with us.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>

Conflicts:

	ChangeLog
	plugins/gadu_protocol/services/gadu-chat-state-service.cpp
	plugins/gadu_protocol/services/gadu-roster-service.cpp

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>