Make use of Qt's safer string.h functions implementations
Now MSVC will not tell us that strdup() is deprecated :P.

BTW, did you know that decent compilers have built-in functions for
strlen() called for C-string literals? So typing 7 instead of
strlen("file://") is not any faster but only more error-prone. And
qstrlen() can make use of it since it's inlined.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>