notify: Fix ScreenModeChecker taking 100% CPU
It was calling a virtual method in its c-tor, and thus (1) never
setting interval on the timer and (2) never connecting it to the slot.
Because of (1) the timer were taking all available resources and because
of (2) the feature was not working at all.

Also free more resources in disable() method.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>