protocol, jabber: Fix immediate effect of accepting untrusted certificate
First problem was in JabberConnectionService where
CertificateErrorWindow was working asynchronously to the connection.
It was supposed to call specific slot which would cause reconnect but
such slot was missing. Protocol::reconnect() was added for this purpose.

Second problem was in MiniClient (used when registering new XMPP
account) which is designed to do it synchronously. It was fixed by
adding an option to handle untrusted certificates synchronously.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>