spellchecker: Switch to using C enchant API
The C++ enchant API isn't any better than the C API, it actually only
gives us overhead. And we did not even have correct implementation of
it, as enchant::Dict::check() could throw and we did not hadle it.

Also, now we don't use exceptions anywhere in Kadu, so I'll be able to
disable them.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>