build: Default to hidden visibility
The main benefit here is that we will be able to early catch issues that
will always result in errors on Windows. Added benefit is that our
shared objects won't be exporting unnecessary symbols, thus reducing
work needed to be done by the dynamic linker -- that's why a lot of
projects, including Qt and KDE, already use -fvisibility=hidden on
Linux.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>