crash fix: Revert "jabber: Fix connecting to server"
Now that we have a proper fix in libiris, we don't need this hack. It
caused crashes for me in case XmppClient was inactive at the time of
JabberConnectionService::cleanUp() call, so we didn't call close() on
it, and later it was trying to use an already deleted Stream object
through a danling poiter.

This reverts commit 438a27af0359150c1f7305c5e8b29da081fc95b6.

Signed-off-by: Bartosz Brachaczek <b.brachaczek@gmail.com>