contacts: do not mark just added contact as dirty, that confuses gadu
That was a mistake - non blocking event when new contact is created and
added to manager. Code that marks contacts to add is still in gui part
as gui part calls setOwnerBuddy and then
Roster::instance()->addContact(contact);

Proper solution would be to remove gadu contacts list concept and merge
it into its roster.

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>